Correct VxNewAligned parameter type and alignment validation
Change align parameter from int to size_t for type correctness. Add
proper handling for zero alignment by defaulting to sizeof(void*).
Ensure alignment validation runs before platform-specific allocation
for consistent behavior across MSVC and other platforms.

+++ include/VxMemory.h
@@ -45,7 +45,7 @@ VX_EXPORT void mydeletearray(void *a);
  * @param align The alignment boundary. Must be a power of two.
  * @return A pointer to the aligned allocated memory, or null on failure.
  */
-VX_EXPORT void *VxNewAligned(size_t size, int align);
+VX_EXPORT void *VxNewAligned(size_t size, size_t align);
 
 /**
  * @brief Deallocates a block of memory that was allocated with VxNewAligned.
